+++ src/apis/hooks/index.ts
@@ -10,9 +10,7 @@ export { default as useCreateSurvey } from './useCreateSurvey'
 export { default as useSignUp } from './useSignUp'
 export { default as useCheckDuplicatedName } from './useCheckDuplicatedName'
 export { default as useCheckDuplicatedEmail } from './useCheckDuplicatedEmail'
-export { default as useGetInvitedReview } from './useGetInvitedReview'
-export { default as useGetCreatedReview } from './useGetCreatedReview'
-export { default as useGetReceivedReview } from './useGetReceivedReview'
+export { default as useGetAllReviews } from './useGetAllReviews'
 export { default as useGetReviewResult } from './useGetReviewResult'
 export { default as useGetResponseForm } from './useGetResponseForm'
 export { default as useGetAllUser } from './useGetAllUser'
